rebuild the GUI on top of $reschedule:
it picks a delayed goals and runs a user given goal in its
context passing the frozen key.  Very primitive but sufficient.

The names of the hv could be typed and parsed, but the lexer is
not utf8 aware, hence a naming facility is added: after intro (i)
one has to name the hyp with a string (n "foo") and can then reuse
(hyp "foo") as a term since it whd to its corresponding hv and it
types as the corresponding hv (this uses the ability to insert clauses
in place, since one has to backpatch whd).