Lists parsed in a reasonable way.
NOTE: camlp5 parsers are not tail recursive.
Every new parsing level makes the call stack
higher. Starting with this commit, the Grundlagen
now require a bigger stack :-(.
NOTE: the Grundlagen do not need all of pervasives.elpi
If one parsing level is commented out (e.g. 160),
the need for a larger stack disappears.

[1,2,3] is parsed as 1::2::3::nil
[(1,2),3] is parsed as (1,2)::3::nil
[ a :- b ] is rejected as a parsing error

The elements of a list are parsed at precedence
level 120, the first one after 110 in pervasives,
which is the level of ",". Teyjus behaves differently
in some situations.