A symbol can now be declared twice, etc.
It is allowed by Teyjus, e.g. a fixity declaration can be put
both in the .mod and in the .sig file.

PARTIAL COMMIT: I do not try yet to detect this situation to
avoid adding a duplicate parsing rule. A camlp5 warning is
printed when that happens (e.g. in pcf test).