WIP: implementation of slim_deadlock.pdf
It uses semi-unification to implement polymorphic recursion, because
most contracts seems to require it.

I also commit a paper that explains how to implement Milner-Mycroft
polymorphism using semi-unification.

IDEA: can we revert the process and replace semi-unification with some
form of explicit management of polymorphism? And then solve the issue
for Hindley-Milner as well? Hint: always quantify a schema over a list
whose size will be bound at the end of inference.