Implementation of local.
local declares variables that are sigma-bound until the end of the scope.
For the time being, however, the scope extends to the end of the code.
I.e. when accumulating a file, the file is included verbatim, and therefore
the scope of locals declared in that file is not terminated.