Bug fixed in delay
$delay performs the opposite action of clausify
to build the delayed sequent. In particular, it
should turn back "p A" into (pi A \ p A). This was
not done and it is done now.

NOTE: we wrote a function that is very similar to
lift_pat that is used in constraint propagation.
Probably the two codes can be unified.