Emergency commit: $delay disabled
After my last commits the behaviour of some tactics has improved, but
the overall behaviour of itaut is worst. It never completes now on our
examples.

This emergency commit turns off type-checking in the $delay case.
I will restore it when I find a fix. In the meantime, itaut can be used.