choose (axiom of choice) is necessary to implement pairs
NOTE: the only "suspect" computational construction remains rec that should
magically return the canonical inhabitant (choose ' B) of B in place of
diverging. We know that that is not possible, but we also know that we can
reason on rec only in the cases it converges. See also Chargueraud idea
on making a rec total (that still requires < to be well_founded, and
moreover decidable as well).