hyps in delayed goal are now pairs depth*t
Before they were already hmoved to the depth of the conclusion.
That operation is very expensive and required only during printing or
constraint propagation. Moreover, in case of printing the relocation is
not necessary and in case of propagation the hypotheses were hmoved again
to yet another (greater) depth.

Allowing $delay in HOL increases the time by 1/3, but that's necessary
for logical soundness.

NOTE: the reason for the increased time is that we do A LOT of type-checking
that is NOT done in ML (HOL Light) because there term is a private type
that basically caches the information that some subterms are already well
typed. This seems a problem unrelated to $delay, but that was hidden by
not delaying.

NOTE: it is true that we have duplicates in the constraints. However, our
current efficient constraint propagation mechanism (in the branch functions)
still introduces an overall slowdown that makes everything slower in the
end. To be further optimized.