complete the "meta level" with $is_name, $llam_unif and $assign
llam_unif (when delaying HO problems is off) is like == in caml, we can
try that before our unification, but won't work most of the times.

$is_name is
  rule (nabla x\ (x T)) (nabla y\ (y T1)) > x ~ y <=> (unif T T1)
that is useful to implement
  unif X X
without adding (unif x x => ...) at all pi x\, but preventing
unif X X from doing things we don't want.
Should we have a syntax like ?? for it?

$assign is the rest of (?? K).