Revert pivoting (means quick uvar->uvar delayed goal propagation)
Regular assignment don't change r.rest (delayed goals) so
making r.contents := dummy restored the old state (the trail
also contains actions to re-add/delete delayed goals from the
store).

UVar := UVar typically makes a delayed goal to be resumed and
re-delayed.  In this case the new UVar "inherits" the delayed
goal.

Pivoting wanted to shortcut this last scenario by just
patching the r.rest list.  This requires to un-pivot when
the uvar is un-assigned, but we had no way to only un-pivot
uvar assigned in the fast path and not the ones assigned in the
regular path.