quotations: strings in {{ }} delegated to 3rd party
This notational artifact is useful to use the parser of
the object language (i.e. the parser of the system in
which elpi is embedded) in order to construct a term
of the object language.

See coq-elpi as an example of this.  Appetizer
  {{ nat -> bool }}
unfolds at program "compilation" time to
  prod _ (indt "...nat") x\ indt "...bool"

Anti quotations are also possible
  prod "x" t x\ {{ nat -> lp:x * bool }}
unfolds to
  prod "x" t x\ prod _ (indt "...nat") y\
    app [indt "...prod", x, indt "...bool"]
Note the x is bound in elpi and used inside the quotation
via the "lp:" anti-quotation.
Binder in one world are only accessible in the same world.

The API is based on an extensible state that is used while
compiling the term (each quotation can store data in such state,
but only access its own data).