hack to make propagation `Spread work
hmove & company have a loose API: ~from must be the level at which
the term lives, not any level.  Indeed a comment warns about silly
restrictions happening if this is not the case.

Lifting patterns cannot just call hmove, we hack by first forcing a complete
dereferencing of the term, and the calling hmove with the "wrong" arguments.

I think we really lack a "lift" function that does not restrict but
dereferences uvars...