Bug avoided and not fully understood.
It can happen that a constraint is delayed on two equal metavariables.
Example: $delay foo [A;A]. Later this determines that the constraint is
added twice to CS.to_resume and thus remove is called twice on it, failing.
Implemented workaround: add to CS.to_resume only if not there yet.

Other workaround that does NOT work (why???): put a sort_uniq
on blockers into add/remove. Everything fails early and I don't understand
why.