freezing in CHR rules done in an intelligible way
Instead of using $match_frozen to destruct frozen
UVars, we represent them as (uvar key [args]), a built in
applicative node.

Thaw is still very incomplete, but at least we can document
what freezing does. Also the (?? KEY ARGS) is really just
the same thing (KEY being flexible), so the parallel between
the meta-runtime and the ?? we do on paper is really there.