native support for wildcard _
New "Discard" term that unifies with everything without computing
the solution. "_" and "_Whatever" are compiled to Discard.

Left aside the potential speedup, the wildcard semantics is
particularly useful when writing deep patterns for dependently
typed object languages.  For example

  search (lam _ x\ lam _ y\ F x y) :- print "found it"

fails if the type of the second lambda node depends on x.  I.e. one
used to be obliged to write the silly

  search (lam _ x\ lam (_ x) y\ F x y) :- print "found it"

This is too error prone (and verbose).