FFI: contextual conversion of arguments
When converting arguments of a built in predicate depend on the
context of the program (hypotheses and constraints) we can now
provide a conversion function for the context that is run once
and forall and whose output is the passed to the conversion functions
of the arguments.

Limitation: for each built in only one context conversion function is
allowed.