[fix] shorten for builtins
When we apply the substitution we adjust, depending on the status
of the substituted constant.

Another approach would be to recognize builtins at the very end of the
compiation chain, but it would require to control all accesses to ToDBL
(eg term_of_ast) and apply there this last phase.