[ffi] InOut takes an ioarg
In this was it is possible to inspect node by node the term
passed as InOut. The OCaml type discipline does not let you
impose it, but each Conversion.t passed to InOut should interleave
a type and ioarg, for example "int option" should be
"int ioarg option ioarg". In this way the argument can be
any of: X, none, some X, some 3.