fixes for camlp5 8.00
Four sorts of fixes:

(1) Pervasives -> Stdlib (to allow clean compile with ocaml 4.10.0)

Maybe these should be removed?

(2) increase test timeout in Makefile (so tests can all pass)

Maybe remove?

(3) the type of Grammar rule has changed (there is now a "digest"
string to allow comparing actions for equality -- this was needed
because of bugs in left-factorization with such equality)

(4) the interface to Plexing has changed (instead of a function for
returning locations, there's a Location.t type -- needed to support
doc-comment extraction)

Both #3, #4 are superficial.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch camlp5-8.00-patches
# Changes to be committed:
#	modified:   Makefile
#	modified:   elpi.opam
#	modified:   src/.ppcache/API.mli
#	modified:   src/.ppcache/ast.ml
#	modified:   src/.ppcache/compiler.ml
#	modified:   src/.ppcache/data.ml
#	modified:   src/.ppcache/runtime_trace_off.ml
#	modified:   src/.ppcache/runtime_trace_on.ml
#	modified:   src/ast.ml
#	modified:   src/builtin.ml
#	modified:   src/parser.ml
#	modified:   src/ptmap.ml
#	modified:   src/runtime.ml
#	modified:   src/util.ml
#