[determinacy] determine type of variable from typeassignment
It is not known a priori the functionality of a fresh unification variable,
for example:
the variable `X` in
`t1 :- pi x\ u (y\ X x y), std.assert! (X 1 2 = 2) "bug".`
cannot be deduced from `u` without typechecking.
We reuse the type assignment given by the typechecker
to the term `X x y` to deduce the functionality of `X`

New:
- typeassignment translation to functionality
- get_functionality_of_term instead returning none if a variable is not
  in the env, returns the conversion of its type assignment to functionality
-